+++ tests/ops_test.py
@@ -29,3 +29,21 @@ def test_nop():
 
     m = poptorch.inferenceModel(Model())
     m(torch.randn(5))
+
+
+def test_name_scope():
+    class Model(torch.nn.Module):
+        def forward(self, x, y):
+            with poptorch.NameScope("NameScope"):
+                return x + y
+
+    model = Model()
+    poptorch_model = poptorch.inferenceModel(model)
+
+    torch.manual_seed(42)
+    x = torch.randn(10, 10)
+    y = torch.randn(10, 10)
+    poptorch_model(x, y)
+
+    ir = poptorch_model._debugGetPopartIR()  # pylint: disable=protected-access
+    assert ir.find('"name":"NameScope/Add:InPlace"') != -1